Add Admin to my project

+++ pages/api/products/[id].js
@@ -9,7 +9,7 @@ export default async function handler(req, res) {
   } = req;
   const token = cookies.token;
 
-  dbConnect();
+  await dbConnect();
 
   if (method === 'GET') {
     try {
@@ -35,9 +35,9 @@ export default async function handler(req, res) {
   }
 
   if (method === 'DELETE') {
-    //if (!token || token !== process.env.token) {
-    // return res.status(401).json('Not authenticated!');
-    //}
+    if (!token || token !== process.env.token) {
+      return res.status(401).json('Not authenticated!');
+    }
     try {
       await Product.findByIdAndDelete(id);
       res.status(200).json('The product has been deleted!');
